moved ShowTravelPlan to a seperate file
- also cleaned the code a little bit and removed the unnecessary use state hook
