+++ pages/index.js
@@ -4,7 +4,7 @@ export default function Home() {
   return (
     <div>
       <h1>Home</h1>
-      <Map></Map>
+      <Map/>
     </div>
   );
 }
